requirements.txt
asgiref==3.8.1
astroid==3.1.0
dill==0.3.8
Django==5.0.3
django-extensions==3.2.3
graphql-core==3.2.3
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-stdnum==1.20
pytz==2024.1
PyYAML==6.0.1
ruff==0.4.2
six==1.16.0
sqlparse==0.4.4
strawberry-graphql==0.227.2
strawberry-graphql-django==0.39.1
tomlkit==0.12.4
typing_extensions==4.11.0
uritemplate==4.1.1
